pyproject.toml
[project]
name = "cts-email-mcp"
version = "1.1.0"
description = "Cantops Enterprise Email & Document MCP Server"
dependencies = [
    "mcp[server]",
    "beautifulsoup4",
    "pydantic-settings",
    "python-pptx",
    "python-docx",
    "openpyxl",
    "pypdf"
]
requires-python = ">=3.10"

[project.scripts]
cts-email-server = "cantops_email.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cantops_email"]
